backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
psycopg2-binary==2.9.10
